Changes Routes folder to a single Router file with all routes connected do Server.ts.
